之前不能调整图像对比度，在im_corr里面先变成uint16再改为single才行

+++ IsletPipeline/im_corr.m
@@ -2,8 +2,8 @@ function [dx,dy,zero_mark] = im_corr(im1,im2,adjust)
 %IM_CORR 此处显示有关此函数的摘要
 %   此处显示详细说明
 if(strcmp(adjust,'adjust'))
-       im1 = imadjust(im1);
-       im2 = imadjust(im2);
+       im1 = single(imadjust(uint16(im1)));
+       im2 = single(imadjust(uint16(im2)));
 end
 im1_fft = fft2(im1);
 im2_fft = fft2(im2);
